Add files via upload
This repository includes an exploratory data analysis project that seeks to understand the determinants of earnings variation, in particular gender, experience, and other socio-demographic characteristics. It uses visual representations and statistical methods to consolidate how pay is apportioned across the various divisions of the organizations and attempts to make sense of pay inequity. The project involves Python programming, data transformation methods, and graphs – all displaying characteristics and correlations – which help in understanding the salary aspects under investigation in an empirical manner.